requirements.txt
django>=3.1.6,<3.2.0
djangorestframework>=3.12.2,<3.20.0
django-simple-history==2.12.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0